requirements.txt
aiogram==3.23.0
gspread==6.2.1
oauth2client==4.1.3
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
python-dotenv==1.2.1
pandas==2.3.3